Override font family to fix text overflow
The previous attempt to fix #3 didn't take into account the fact that
some users are seeing a completely different font than most other users.
It turns out that boardgamearena specifies a "Roboto" font for the body,
but never imports that font (it does however import a font called
'robotoregular', so this might be a typo situation). It also specifies
Arial as a fallback font, so that's what most people see. If a user
happens to have a font called Roboto installed on their system however,
that will take precedence and result in unpredictable text overflow.
This overrides the font-family (just for the card list) to use Arial so
that users will get a consistent experience.

This closes #3